CtrlCore: Fixed issue with WM_HOTKEY (thanks Zbych!)
git-svn-id: svn://ultimatepp.org/upp/trunk@6614 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlCore/Win32Wnd.cpp
@@ -754,7 +754,7 @@ void Ctrl::UnregisterSystemHotKey(int id)
 void Ctrl::sProcessMSG(MSG& msg)
 {
 	if (msg.message == WM_HOTKEY) {
-		if(msg.wParam >= 0 && (int)msg.wParam < Ctrl::hotkey.GetCount())
+		if((int)msg.wParam >= 0 && (int)msg.wParam < Ctrl::hotkey.GetCount())
 			Ctrl::hotkey[(int)msg.wParam]();
 		return;
 	}
